+++ hbw/app/javascript/packs/hbw/widget.js.jsx
@@ -136,7 +136,7 @@ modulejs.define(
 
       setWidgetContainer = (container) => {
         this.$widgetContainer = [container];
-      }
+      };
     }
 
     return HBW;